+++ changelog.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Unique ability stat for Frost
 * Unique ability stat for Caveira
 * DP27 weapon
+* Past seasons stats for `getRanks`
 
 ### Changed
 * Unique ability for Tachanka